feature: 조직 - 유저 연결 엔티티 추가 및 서비스 메소드 (추가/삭제) 추가

+++ src/main/kotlin/jjfactory/organization/organization/team/TeamServiceImpl.kt
@@ -1,5 +1,6 @@
 package jjfactory.organization.organization.team
 
+import jjfactory.organization.user.UserRepository
 import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -8,15 +9,40 @@ import org.springframework.transaction.annotation.Transactional
 @Transactional
 @Service
 class TeamServiceImpl(
-    private val teamRepository: TeamRepository
+    private val teamRepository: TeamRepository,
+    private val userRepository: UserRepository,
+    private val teamUserRepository: TeamUserRepository
 ) : TeamService {
     override fun createTeam(request: TeamDto.CreateRequest): Long {
         return teamRepository.save(request.toEntity()).id!!
     }
 
     override fun delete(id: Long) {
-        teamRepository.findByIdOrNull(id) ?: throw NotFoundException()
+        val team = teamRepository.findByIdOrNull(id) ?: throw NotFoundException()
+        val teamUsers = teamUserRepository.findAllByTeamId(teamId = team.id!!)
+
+        if (teamUsers.isNotEmpty()){
+            //FIXME 익셉션 타입변경
+            throw IllegalArgumentException("조직원이 존재하는 팀은 삭제 불가")
+        }
+
         teamRepository.deleteById(id)
+    }
+
+    override fun addUsersToTeam(requests: List<TeamDto.AddUserRequest>, teamId: Long){
+        val team = teamRepository.findByIdOrNull(teamId) ?: throw NotFoundException()
+
+        requests.forEach {
+            val user = userRepository.findByIdOrNull(teamId) ?: throw NotFoundException()
+
+            val initTeamUser = TeamUser(
+                teamId = team.id!!,
+                userId = user.id!!,
+                position = it.position
+            )
+
+            teamUserRepository.save(initTeamUser)
+        }
     }
 
     override fun update(id: Long, request: TeamDto.UpdateRequest): Long {
